changelog service: readme and request refactoring
