+++ src/Changelog/Query/Requests/getLogsOfCourseRequest.php
@@ -12,10 +12,7 @@ use ILIAS\Changelog\Query\Requests\Filters\getLogsOfUserFilter;
  *
  * @author Theodor Truffer <tt@studer-raimann.ch>
  */
-class getLogsOfCourseRequest {
-
-	const ORDER_ASCENDING = 'ASC';
-	const ORDER_DESCENDING = 'DESC';
+class getLogsOfCourseRequest extends Request {
 
 	/**
 	 * @var int
@@ -26,24 +23,6 @@ class getLogsOfCourseRequest {
 	 * @var getLogsOfUserFilter
 	 */
 	protected $filter;
-	/**
-	 * @var int
-	 */
-	protected $limit = 0;
-	/**
-	 * @var int
-	 */
-	protected $offset = 0;
-
-	/**
-	 * @var string
-	 */
-	protected $orderBy = 'timestamp';
-
-	/**
-	 * @var string
-	 */
-	protected $orderDirection = 'ASC';
 
 	/**
 	 * getLogsOfUsersRequest constructor.
@@ -61,33 +40,6 @@ class getLogsOfCourseRequest {
 		$this->filter = $getLogsOfUsersFilter;
 	}
 
-	/**
-	 * @return int
-	 */
-	public function getLimit(): int {
-		return $this->limit;
-	}
-
-	/**
-	 * @param int $limit
-	 */
-	public function setLimit(int $limit) {
-		$this->limit = $limit;
-	}
-
-	/**
-	 * @return int
-	 */
-	public function getOffset(): int {
-		return $this->offset;
-	}
-
-	/**
-	 * @param int $offset
-	 */
-	public function setOffset(int $offset) {
-		$this->offset = $offset;
-	}
 
 	/**
 	 * @return int
@@ -103,31 +55,4 @@ class getLogsOfCourseRequest {
 		return $this->filter;
 	}
 
-	/**
-	 * @return string
-	 */
-	public function getOrderBy(): string {
-		return $this->orderBy;
-	}
-
-	/**
-	 * @param string $orderBy
-	 */
-	public function setOrderBy(string $orderBy) {
-		$this->orderBy = $orderBy;
-	}
-
-	/**
-	 * @return string
-	 */
-	public function getOrderDirection(): string {
-		return $this->orderDirection;
-	}
-
-	/**
-	 * @param string $orderDirection
-	 */
-	public function setOrderDirection(string $orderDirection) {
-		$this->orderDirection = $orderDirection;
-	}
 }
